+++ src/main/java/pl/lodz/p/it/ssbd2023/ssbd03/mok/Manager.java
@@ -6,7 +6,7 @@ public class Manager extends Account {
 
     private String license;
 
-    public Manager(int id, String email, String password, Boolean isEnable, LocalDateTime registerDate, String language, LoginData loginData, PersonalData personalData, String license) {
+    public Manager(Long id, String email, String password, Boolean isEnable, LocalDateTime registerDate, String language, LoginData loginData, PersonalData personalData, String license) {
         super(id, email, password, isEnable, registerDate, language, loginData, personalData);
         this.license=license;
     }
